make an incremental crate
for now, this houses `svh` and the code to check `assert_dep_graph` is
sane
