+++ src/librustc/lib.rs
@@ -100,6 +100,7 @@ pub mod middle {
     pub mod recursion_limit;
     pub mod resolve_lifetime;
     pub mod stability;
+    pub mod svh;
     pub mod weak_lang_items;
 }
