+++ src/librustc_incremental/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+authors = ["The Rust Project Developers"]
+name = "rustc_serialize"
+version = "0.0.0"
+
+[lib]
+name = "rustc_serialize"
+path = "lib.rs"
+crate-type = ["dylib"]
+
+[dependencies]
+graphviz = { path = "../libgraphviz" }
+rbml = { path = "../librbml" }
+rustc = { path = "../librustc" }
+rustc_data_structures = { path = "../librustc_data_structures" }
+rustc_front = { path = "../librustc_front" }
+serialize = { path = "../libserialize" }
